refactor(renderer): extract visual-enhancement resolution from index.ts
Moves resolveVisualEnhancement, the ResolvedVisualEnhancement type, and the
per-frame merged state into a new VisualEnhancementResolver collaborator
(visual-enhancement.ts). The state was read/written only inside that one
method with a single caller in renderFrame, so it lifts out cleanly per
the seam-extraction pattern in issue #2425 (render-degradation.ts).

Pure move, behaviour-preserving: no public API change (both were private
to Renderer), no test/getDiagnostics/destroy references to the moved
state. index.ts drops from 3946 to 3894 lines.

Refs #2425
